# src/docs/swagger.yaml
basePath: /
definitions:
  entity.CreateUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  entity.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: 192.168.2.117:3000
info:
  contact: {}
  description: An example of gin
  license:
    name: MIT
  title: go-docker API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      - application/x-json-stream
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - Root
  /create_user:
    post:
      consumes:
      - application/json
      - application/x-json-stream
      parameters:
      - description: CreateUserRequest
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/entity.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - Root
  /login:
    post:
      consumes:
      - application/json
      - application/x-json-stream
      parameters:
      - description: LoginRequest
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/entity.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - Root
  /logout:
    get:
      consumes:
      - application/json
      - application/x-json-stream
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - Root
  /menu/top:
    get:
      consumes:
      - application/json
      - application/x-json-stream
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - Root
  /ping:
    get:
      consumes:
      - application/json
      - application/x-json-stream
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - Root
swagger: "2.0"